# api/openapi.yml
openapi: 3.1.0
info:
  title: EMS API
  description: |
    Energy Management System API 規範

    本系統包含以下 API 端點：
    - **Simulator API** (port 8001): 模擬電表控制
    - **Query API** (port 3001): 歷史資料查詢 (PostgREST)
    - **Grafana API** (port 3000): 告警與儀表板管理
    - **Device Service API** (port 8002): Device Registry 與 AI 自動分類 (PRD-0003)
    - **Grafana Dashboard**: 已新增 KC 工廠環境與設備狀態 panels

    注意：device-service 另有 MCP 端點 (127.0.0.1:8766)，走 MCP Streamable-HTTP 協定
    而非 REST，因此不在本 OpenAPI 規範內；詳見 doc/operations/容器速查表 §2-14。
  version: 1.5.0
  contact:
    name: EMS Development Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8001
    description: Simulator API (模擬電表控制)
  - url: http://localhost:3001
    description: Query API - PostgREST (歷史資料查詢)
  - url: http://localhost:3000
    description: Grafana API (告警與儀表板)
  - url: http://localhost:8002
    description: Device Service API (Device Registry + AI 自動分類, PRD-0003)
...
paths:
  /api/alarms/demo:
    servers: [{url: 'http://localhost:8003'}]
    get:
      tags: [BFF Demo Alarms]
      summary: Telegram 示範通知設定狀態與最近十筆成功紀錄
      operationId: getDemoAlarmStatus
      description: OPS session required. No query parameters. No external send. In-memory records clear on BFF restart.
      security: [{bffSession: []}]
      responses:
        '200':
          description: Notification readiness and recent successful acknowledgements
          content:
            application/json:
              schema:
                type: object
                required: [configured, cooldown_seconds, recent]
                properties:
                  configured: {type: boolean}
                  cooldown_seconds: {type: integer, minimum: 0}
                  recent:
                    type: array
                    maxItems: 10
                    items: {$ref: '#/components/schemas/DemoAlarmDelivery'}
        '401': {description: Session absent or expired}
        '403': {description: OPS role required}
        '422': {description: Query parameters are not supported}
    post:
      tags: [BFF Demo Alarms]
      summary: 發送固定對象與固定 DEMO 文字的 Telegram 示範警報
      operationId: triggerDemoAlarm
      description: |
        OPS session and allowlisted Origin required. No query parameters.
        Server-side fixed bot, recipient and DEMO message; no device control.
        Global 30-second default cooldown. Same UUID is deduplicated for 10 minutes
        including failures (64-entry cap); no automatic retry. Single-process,
        in-memory state resets on restart. Success means Telegram accepted, not read.
      security: [{bffSession: []}]
      parameters:
        - name: Origin
          in: header
          required: true
          schema: {type: string, format: uri}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [request_id]
              properties:
                request_id:
                  type: string
                  format: uuid
                  description: UUID version 4; reuse only for the same logical send attempt
      responses:
        '200':
          description: Confirmed Telegram acknowledgement or replay of the same successful UUID
          content:
            application/json:
              schema:
                allOf:
                  - {$ref: '#/components/schemas/DemoAlarmDelivery'}
                  - type: object
                    required: [cooldown_seconds]
                    properties:
                      cooldown_seconds: {type: integer, minimum: 0}
        '401': {description: Session absent or expired}
        '403': {description: OPS role or Origin check failed}
        '409': {description: Same UUID is still in progress (demo_alarm_in_progress)}
        '422': {description: Invalid UUID4 or unsupported fields/query}
        '429':
          description: Global cooldown or bounded dedupe cache is full (demo_alarm_cooldown)
          headers:
            Retry-After:
              schema: {type: integer, minimum: 1}
              description: Seconds before a new attempt is permitted
        '502':
          description: telegram_delivery_failed or telegram_delivery_unknown; no automatic resend
        '503': {description: Telegram credentials missing (telegram_not_configured)}
  /api/devices/{device_id}/history:
    servers: [{url: 'http://localhost:8003'}]
    get:
      tags: [BFF History]
      summary: 完整指定區間的設備趨勢彙整
      operationId: getDeviceHistory
      description: |
        OPS session required. Read-only and idempotent; retries safe.
        Half-open [since, until), timezone required, maximum 7 days.
        Unknown/repeated query parameters return 422. No new rate limiter.
        Server resolves domain through gateway_id and calls internal read-only
        api.measurement_history RPC (migration 017); never accepts browser SQL.
        Numeric value is mean; last is bucket-end sample. Boolean signals encode
        0/1; min != max means transitions within the bucket, not an exact timeline.
        Empty buckets/signals are omitted; missing data must not be filled with zero.
      security: [{bffSession: []}]
      parameters:
        - {$ref: '#/components/parameters/HistoryDeviceId'}
        - {$ref: '#/components/parameters/HistorySince'}
        - {$ref: '#/components/parameters/HistoryUntil'}
        - name: points
          in: query
          schema: {type: integer, minimum: 60, maximum: 1200, default: 600}
      responses:
        '200':
          description: Bounded bucket series for the entire requested interval
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeviceHistory'}
        '401': {description: Session absent or expired}
        '403': {description: OPS role required}
        '404': {description: Device or measurement domain missing}
        '422': {description: Invalid range or parameters}
        '502': {description: Upstream unavailable}
        '503': {description: OPS channel not configured}
  /api/devices/{device_id}/records:
    servers: [{url: 'http://localhost:8003'}]
    get:
      tags: [BFF History]
      summary: 查詢指定區間的原始量測紀錄
      operationId: getDeviceRecords
      description: |
        OPS-only, read-only, idempotent; same 7-day timezone-aware half-open range.
        Ordered by time descending and full row JSON as deterministic tie-break.
        Returns a fixed-time-window page, not a transaction snapshot; late arrivals
        may shift offsets. limit+1 lookahead determines has_more. Existing samples
        are unmodified; fields may be null for sparse factory rows.
        Internal RPC api.measurement_records, via BFF only for browser clients.
      security: [{bffSession: []}]
      parameters:
        - {$ref: '#/components/parameters/HistoryDeviceId'}
        - {$ref: '#/components/parameters/HistorySince'}
        - {$ref: '#/components/parameters/HistoryUntil'}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 1000, default: 100}
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, maximum: 1000000, default: 0}
      responses:
        '200':
          description: Raw page; no total count
          content:
            application/json:
              schema:
                type: object
                required: [rows, has_more]
                properties:
                  has_more: {type: boolean}
                  rows:
                    type: array
                    maxItems: 1000
                    items:
                      type: object
                      required: [time, device_id]
                      properties:
                        time: {type: string, format: date-time}
                        device_id: {type: string}
                      additionalProperties: true
        '401': {description: Session absent or expired}
        '403': {description: OPS role required}
        '404': {description: Device or measurement domain missing}
        '422': {description: Invalid range or pagination}
        '502': {description: Upstream unavailable}
        '503': {description: OPS channel not configured}

  # ============================================================
  # Simulator API (port 8001)
  # ============================================================
...
  /electricity_measurements:
    get:
      tags:
        - Query
      summary: 查詢歷史電力量測資料
      description: |
        使用 PostgREST 語法查詢 TimescaleDB 中的歷史電力量測資料（electricity_measurements 表）。

        **支援的運算子：**
        - `eq.` 等於
        - `gt.` 大於
        - `gte.` 大於等於
        - `lt.` 小於
        - `lte.` 小於等於
        - `neq.` 不等於

        **範例：**
        - `?power_kw=gt.50` - 功率大於 50kW
        - `?time=gte.2026-04-23T00:00:00` - 時間大於等於指定時間
      operationId: getMeasurements
      servers:
        - url: http://localhost:3001
      parameters:
        - name: select
          in: query
          description: 選擇要回傳的欄位 (逗號分隔)
          schema:
            type: string
          example: time,power_kw,voltage
        - name: device_id
          in: query
          description: 依設備 ID 過濾 (使用 eq.xxx 格式)
          schema:
            type: string
          example: eq.sim-001
        - name: time
          in: query
          description: 依時間過濾 (使用 gte.xxx 或 lte.xxx 格式)
          schema:
            type: string
          example: gte.2026-04-23T00:00:00
        - name: power_kw
          in: query
          description: 依功率過濾 (使用 gt.xxx, lt.xxx 等格式)
          schema:
            type: string
          example: gt.50
        - name: voltage
          in: query
          description: 依電壓過濾
          schema:
            type: string
        - name: current
          in: query
          description: 依電流過濾
          schema:
            type: string
        - name: order
          in: query
          description: 排序欄位 (欄位名.asc 或 欄位名.desc)
          schema:
            type: string
            default: time.desc
          example: time.desc
        - name: limit
          in: query
          description: 限制回傳筆數
          schema:
            type: integer
            minimum: 1
            maximum: 10000
            default: 100
          example: 10
        - name: offset
          in: query
          description: 跳過前 N 筆
          schema:
            type: integer
            minimum: 0
          example: 0
      responses:
        '200':
          description: 量測資料列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Measurement'
              example:
                - time: "2026-04-23T10:30:00Z"
                  device_id: sim-001
                  voltage: 381.2
                  current: 105.3
                  power_kw: 58.7
                  energy_kwh: 1234.56
        '400':
          description: 查詢參數錯誤
...
  /devices:
    post:
      tags: [Device Service]
      summary: 建立 candidate 裝置（OPS）
      operationId: createDevice
      security: [{ deviceServiceApiKey: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/DeviceCreate' } } }
      responses:
        '201': { description: 已建立, content: { application/json: { schema: { $ref: '#/components/schemas/DeviceOut' } } } }
        '401': { description: 缺/錯 X-API-Key }
        '403': { description: 通道不允許 }
        '409': { description: 裝置已存在 }
    get:
      tags: [Device Service]
      summary: 列出裝置（OPS；可依 status / stale / type 過濾，支援分頁與排序）
      operationId: listDevices
      security: [{ deviceServiceApiKey: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: stale, in: query, schema: { type: boolean } }
        - { name: type, in: query, schema: { type: string }, description: '依 device_type 過濾' }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 }, description: '分頁筆數上限（不給則回全部）' }
        - { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: sort, in: query, schema: { type: string, enum: [device_id, device_type, status, ai_confidence, created_at, updated_at, last_seen_at], default: device_id }, description: '排序欄位 allowlist；NULLS LAST，次序鍵 device_id' }
        - { name: order, in: query, schema: { type: string, enum: [asc, desc], default: asc } }
      responses:
        '200': { description: 裝置清單, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/DeviceOut' } } } } }
        '422': { description: 參數驗證失敗（如 sort 不在 allowlist）}
...
components:
  parameters:
    HistoryDeviceId:
      name: device_id
      in: path
      required: true
      schema: {type: string, pattern: '^[A-Za-z0-9_-]{1,64}$'}
    HistorySince:
      name: since
      in: query
      required: true
      schema: {type: string, format: date-time}
      description: Inclusive start with timezone; total interval at most 7 days
    HistoryUntil:
      name: until
      in: query
      required: true
      schema: {type: string, format: date-time}
      description: Exclusive end with timezone
  securitySchemes:
    bffSession:
      type: apiKey
      in: cookie
      name: ems_bff_session
      description: HttpOnly Secure SameSite Strict opaque session; OPS role required
    basicAuth:
      type: http
      scheme: basic
      description: Grafana admin 帳號密碼 (預設 admin:admin)
    deviceServiceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        device-service 三通道金鑰（FR-310）。OPS 通道用於人工 lifecycle/feedback；
        INGEST 通道供 gateway 寫遙測來源；AI 通道供分類/MCP。金鑰存於 .env
        （OPS_API_KEY / INGEST_API_KEY / AI_API_KEY），缺值該通道 fail-closed。
        AI 金鑰輪替後需 `docker compose up -d device-service-mcp` 重啟（L-3）。
  schemas:
    DemoAlarmDelivery:
      type: object
      required: [status, event_id, sent_at, channel]
      properties:
        status: {type: string, enum: [sent]}
        event_id: {type: string, format: uuid}
        sent_at: {type: string, format: date-time}
        channel: {type: string, enum: [telegram]}
    DeviceHistory:
      type: object
      required: [device_id, since, until, bucket_seconds, series]
      properties:
        device_id: {type: string}
        since: {type: string, format: date-time}
        until: {type: string, format: date-time}
        bucket_seconds: {type: integer, minimum: 1, maximum: 10080}
        series:
          type: array
          maxItems: 7200
          items:
            type: object
            required: [time, signal, value, min, max, last, samples, last_time]
            properties:
              time: {type: string, format: date-time, description: Bucket start anchored to since}
              signal:
                type: string
                enum: [voltage, current, power_kw, energy_kwh, temperature, humidity, motor_speed, pressure, pump_on, valve_open]
              value: {type: number, description: Mean of finite samples}
              min: {type: number}
              max: {type: number}
              last: {type: number, description: Latest finite sample in the bucket}
              samples: {type: integer, minimum: 1}
              last_time: {type: string, format: date-time}

    # ===== Device Service (PRD-0003) — 對齊 services/device-service/device_service/models.py =====
    DeviceCreate:
      type: object
      required: [device_id]
      properties:
        device_id: { type: string, pattern: '^[a-zA-Z0-9_-]{1,64}$', example: sim-002 }
        device_type: { type: string, nullable: true, description: 'electricity / temperature / unknown ...' }
        protocol: { type: string, nullable: true, example: modbus_tcp }
        vendor: { type: string, nullable: true }
        model: { type: string, nullable: true }
        location: { type: string, nullable: true }
        gateway_id: { type: string, nullable: true, description: 來源 gateway, example: ems-gateway }
...
    DeviceOut:
      type: object
      required: [device_id, status]
      properties:
        device_id: { type: string }
        device_type: { type: string, nullable: true }
        status: { type: string, description: 'candidate / confirmed / retired' }
        protocol: { type: string, nullable: true }
        vendor: { type: string, nullable: true }
        model: { type: string, nullable: true }
        location: { type: string, nullable: true }
        gateway_id: { type: string, nullable: true }
        classified_by: { type: string, nullable: true, description: 'ai / human / manual_override / migration_backfill' }
        ai_confidence: { type: number, format: float, nullable: true, minimum: 0, maximum: 1, description: 'AI 分類信心值（0–1；未分類為 null）' }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
        last_seen_at: { type: string, format: date-time, nullable: true }
        confirmed_at: { type: string, format: date-time, nullable: true }
...
    Measurement:
      type: object
      required:
        - time
        - device_id
      properties:
        time:
          type: string
          format: date-time
          description: 量測時間 (UTC)
          example: "2026-04-23T10:30:00Z"
        device_id:
          type: string
          description: 設備 ID
          example: sim-001
        voltage:
          type: number
          format: float
          description: 電壓 (V)
          example: 381.2
        current:
          type: number
          format: float
          description: 電流 (A)
          example: 105.3
        power_kw:
          type: number
          format: float
          description: 功率 (kW)
          example: 58.7
        energy_kwh:
          type: number
          format: float
          description: 累計電量 (kWh)
          example: 1234.56
...